Completely new system of operations on industries aiming to get rid of GameProperties in most cases (and making the project modifiable!).
